167 distributed

+++ mlprogram/entrypoint/train.py
@@ -79,16 +79,16 @@ def create_extensions_manager(n_iter: int, evaluation_interval_iter: int,
         extensions.FailOnNonNumber(),
         trigger=Trigger(evaluation_interval_iter, n_iter)
     )
+    if evaluate is not None:
+        manager.extend(
+            Call(evaluate),
+            trigger=Trigger(evaluation_interval_iter, n_iter),
+        )
     if distributed.is_main_process():
         manager.extend(
             extensions.LogReport(trigger=Trigger(100, n_iter))
         )
         manager.extend(extensions.ProgressBar())
-        if evaluate is not None:
-            manager.extend(
-                Call(evaluate),
-                trigger=Trigger(evaluation_interval_iter, n_iter),
-            )
         manager.extend(
             SaveTopKModel(model_dir, 1, metric, model, maximize=maximize),
             trigger=Trigger(evaluation_interval_iter, n_iter),
@@ -147,38 +147,49 @@ def get_world_process_group(device: torch.device) \
             return distributed.groups["world_gloo"]
 
 
-def all_reduce(model: nn.Module, group: torch.distributed.group):
-    nparams = list(model.named_parameters())
-    nparams.sort(key=lambda x: x[0])  # type: ignore
-    params = [p.grad if p.grad is not None else torch.zeros_like(p.data)
-              for _, p in nparams]
-    size = distributed.size(group)
-    distributed.call(
-        torch.distributed.all_reduce_coalesced,
-        params, group=group
-    )
-    for p in params:
-        p /= size
+def setup_distributed_training(
+    model: nn.Module,
+    loss: nn.Module,
+    group: torch.distributed.group
+):
+    class TrainModule(nn.Module):
+        def __init__(self, model: nn.Module, loss: nn.Module):
+            super().__init__()
+            self.model = model
+            self.loss = loss
+
+        def forward(self, *args, **kwargs):
+            return self.loss(self.model(*args, **kwargs))
+
+    model = TrainModule(model, loss)
+    if group is None:
+        return model
+    else:
+        return ppe.nn.parallel.distributed.DistributedDataParallel(
+            module=model,
+            process_group=group,
+        )
 
 
 def save_results(workspace_dir: str, output_dir: str,
                  model: nn.Module, optimizer: torch.optim.Optimizer) -> None:
-    model_dir = os.path.join(workspace_dir, "model")
-    logger.info("Copy log to output_dir")
-    if os.path.exists(os.path.join(workspace_dir, "log")):
-        os.makedirs(output_dir, exist_ok=True)
-        shutil.copyfile(os.path.join(workspace_dir, "log"),
-                        os.path.join(output_dir, "log.json"))
+    if distributed.is_main_process():
+        model_dir = os.path.join(workspace_dir, "model")
+        logger.info("Copy log to output_dir")
+        if os.path.exists(os.path.join(workspace_dir, "log")):
+            os.makedirs(output_dir, exist_ok=True)
+            shutil.copyfile(os.path.join(workspace_dir, "log"),
+                            os.path.join(output_dir, "log.json"))
 
-    logger.info("Copy models to output_dir")
-    out_model_dir = os.path.join(output_dir, "model")
-    if os.path.exists(out_model_dir):
-        shutil.rmtree(out_model_dir)
-    shutil.copytree(model_dir, out_model_dir)
+        logger.info("Copy models to output_dir")
+        out_model_dir = os.path.join(output_dir, "model")
+        if os.path.exists(out_model_dir):
+            shutil.rmtree(out_model_dir)
+        shutil.copytree(model_dir, out_model_dir)
 
-    logger.info("Dump the last model")
-    torch.save(model.state_dict(), os.path.join(output_dir, "model.pt"))
-    torch.save(optimizer.state_dict(), os.path.join(output_dir, "optimizer.pt"))
+        logger.info("Dump the last model")
+        torch.save(model.state_dict(), os.path.join(output_dir, "model.pt"))
+        torch.save(optimizer.state_dict(), os.path.join(output_dir, "optimizer.pt"))
 
 
 def train_supervised(workspace_dir: str, output_dir: str,
@@ -228,6 +239,8 @@ def train_supervised(workspace_dir: str, output_dir: str,
             evaluate, metric, maximize, threshold,
             workspace_dir)
 
+    train_model = setup_distributed_training(model, loss, group)
+
     logger.info("Start training")
     try:
         while manager.iteration < n_iter:
@@ -241,17 +254,14 @@ def train_supervised(workspace_dir: str, output_dir: str,
                     logger.warning(f"Skip {manager.iteration} th batch")
                     continue
                 with manager.run_iteration():
-                    model.train()
+                    train_model.train()
                     with logger.block("to"):
                         batch.to(device=device)
                     with logger.block("forward"):
-                        output = model(batch)
-                        bloss = loss(output)
+                        bloss = train_model(batch)
                     with logger.block("backward"):
-                        model.zero_grad()
+                        optimizer.zero_grad(set_to_none=True)
                         bloss.backward()
-                    with logger.block("all-reduce"):
-                        all_reduce(model, group)
                     with logger.block("optimizer.step"):
                         optimizer.step()
 
@@ -333,6 +343,8 @@ def train_REINFORCE(input_dir: str, workspace_dir: str, output_dir: str,
             workspace_dir,
             report_metrics=["reward"])
 
+    train_model = setup_distributed_training(model, loss, group)
+
     logger.info("Start training")
     try:
         while manager.iteration < n_iter:
@@ -344,7 +356,7 @@ def train_REINFORCE(input_dir: str, workspace_dir: str, output_dir: str,
                     break
                 # Rollout
                 rollouts = []
-                model.train()
+                train_model.train()
                 with torch.no_grad():
                     for sample in logger.iterable_block("rollout", samples):
                         sample_inputs = sample.clone_without_supervision()
@@ -377,14 +389,11 @@ def train_REINFORCE(input_dir: str, workspace_dir: str, output_dir: str,
                     with logger.block("to"):
                         batch2.to(device)
                     with logger.block("forward"):
-                        model.train()
-                        output = model(batch2)
-                        bloss = loss(output)
+                        train_model.train()
+                        bloss = train_model(batch2)
                     with logger.block("backward"):
-                        model.zero_grad()
+                        optimizer.zero_grad(set_to_none=True)
                         bloss.backward()
-                    with logger.block("all-reduce"):
-                        all_reduce(model, group)
                     with logger.block("optimizer.step"):
                         optimizer.step()
